+++ packages/core/doclet.d.ts
@@ -1,0 +1,4 @@
+export default function (api: {
+  registerDocumentation(section: any, category: any): void;
+  registerSearchProvider(cb: any): void;
+});