+++ packages/core/src/scaffold/index.tsx
@@ -1,6 +1,5 @@
 import type { PiletApi } from '<%- sourceName %>';
-
-const createDoclet = require('@pidoc/core/doclet');
+import createDoclet from '@pidoc/core/doclet';
 
 export function setup(api: PiletApi) {
   createDoclet(api);